requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
cheroot==9.0.0
dacite==1.7.0
idna==3.4
jaraco.functools==3.5.2
more-itertools==9.0.0
Pillow==9.4.0
pokemontcgsdk==3.4.0
pyserial==3.5
python-dotenv==0.21.0
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
six==1.16.0
urllib3==1.26.13
web.py==0.62
